+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.5 (29-05-2020)
+
+- New server index + readme + docs + links + info
+
+## 0.1.4 (28-05-2020)
+
+- Fix? vscode endless loop instalation
+
 ## 0.1.3 (28-05-2020)
 
 - Bug correction for windows users